Move unit tests from distribution testing to implicit priors
