+++ tests/test_distribution.py
@@ -683,46 +683,3 @@ def test_Gaussian_from_sparse_sqrtprec():
 
     assert y_from_dense.logpdf(np.ones(N)) == y_from_sparse.logpdf(np.ones(N))
 
-def test_RegularizedGaussian_default_init():
-    """ Test that the implicit regularized Gaussian is initialized by default with nonnegativity preset """
-
-    x = cuqi.distribution.RegularizedGaussian(np.zeros(5), 1)
-    assert x._preset == "nonnegativity"
-
-def test_RegularizedGaussian_guarding_statements():
-    """ Test that we catch incorrect initialization of RegularizedGaussian """
-
-    # More than 1 argument
-    with pytest.raises(ValueError, match="Only one of "):
-        cuqi.distribution.RegularizedGaussian(np.zeros(5), 1, proximal=lambda s,z: s, constraint="nonnegativity")
-
-    # Proximal
-    with pytest.raises(ValueError, match="Proximal needs to be callable"):
-        cuqi.distribution.RegularizedGaussian(np.zeros(5), 1, proximal=1)
-
-    with pytest.raises(ValueError, match="Proximal should take 2 arguments"):
-        cuqi.distribution.RegularizedGaussian(np.zeros(5), 1, proximal=lambda s: s)
-
-    # Projector
-    with pytest.raises(ValueError, match="Projector needs to be callable"):
-        cuqi.distribution.RegularizedGaussian(np.zeros(5), 1, projector=1)
-
-    with pytest.raises(ValueError, match="Projector should take 1 argument"):
-        cuqi.distribution.RegularizedGaussian(np.zeros(5), 1, projector=lambda s,z: s)
-
-def test_RegularizedGaussian_init_options():
-    """ Test different options for initialization of RegularizedGaussian """
-
-    options = {
-        "constraint": "nonnegativity",
-        "constraint": "NN",
-        "constraint": "non-negativity",
-        "constraint": "non-negative",
-        "constraint": "Box",
-        "regularization": "L1",
-    }
-
-    for key, value in options.items():
-        x = cuqi.distribution.RegularizedGaussian(np.zeros(5), 1, **{key: value}) 
-
-
